refresh and bugs
