--- src/modules/ourHoney/images/index.js
@@ -1,1 +1,0 @@
-export { default as ourHoneySet } from './ourHoneySet.svg';
