+++ src/modules/ourProducts/images/index.js
@@ -1,1 +1,1 @@
-export { default as shopCartSet } from './shopCartSet.svg';
+
