+++ src/shared/components/CustomScrollWrapper/CustomScrollWrapper.scss
@@ -6,7 +6,7 @@
   --os-track-border-radius: 6px;
   --os-handle-border-radius: 6px;
 
-  --os-handle-bg-hover: #ebb41c;
-  --os-handle-bg-active: #ebb41c;
+  --os-handle-bg-hover: #facf5a;
+  --os-handle-bg-active: #facf5a;
 
 }
